requirements.txt
aiohttp==2.3.1
async-timeout==2.0.1
certifi==2018.1.18
chardet==3.0.4
decorator==4.2.1
idna==2.6
idna-ssl==1.0.1
jmespath==0.9.3
multidict==4.1.0
ply==3.11
PyYAML==3.12
requests==2.18.4
six==1.11.0
urllib3==1.22
yarl==1.1.1
